+++ myProject/app/src/main/java/com/blblblbl/myapplication/view/compose_utils/PhotoComposables.kt
@@ -125,3 +125,35 @@ fun PhotoScreen(
         }
     }
 }
+@Composable
+fun StatesUI(items: LazyPagingItems<Photo>){
+    items.apply {
+        when {
+            loadState.refresh is LoadState.Loading -> {
+                LoadingView(modifier = Modifier.fillMaxSize())
+            }
+            loadState.append is LoadState.Loading -> {
+                LoadingItem()
+            }
+            loadState.refresh is LoadState.Error -> {
+                val e = items.loadState.refresh as LoadState.Error
+
+                ErrorItem(
+                    message = e.error.localizedMessage!!,
+                    modifier = Modifier.fillMaxSize(),
+                    onClickRetry = { retry() }
+                )
+
+            }
+            loadState.append is LoadState.Error -> {
+                val e = items.loadState.append as LoadState.Error
+
+                ErrorItem(
+                    message = e.error.localizedMessage!!,
+                    onClickRetry = { retry() }
+                )
+
+            }
+        }
+    }
+}
